move workspace report tools to new module, move tests
